Provide default callbacks to Searcher
The values are None by default, but if a SearchView subclass defines
methods for get_display_fields and/or process_results, they will be
called during data processing.  They are passed the model, the search
configuration for that model, etc.  These might disappear in the future
if they're not useful enough.