Adds Field.choices inspection
Fields that define a choices attribute (regardless of their underlying
type, such as IntegerField) will present the client with text options
(such as "Bob" and "Mary" instead of integers).  If the user enters
such a "label" value, fetch the underlying data from the choices list.

This update makes the search functional, but the presented operator
list will always match the field type (IntegerField) even when the
expected input becomes textual due to this update.